requirements.txt
autopep8==1.7.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
discord-webhook==0.16.3
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pycodestyle==2.9.1
python-dotenv==0.20.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.2.0
toml==0.10.2
urllib3==1.26.11
